Added STM32 support

+++ inc/I2C_HAL.h
@@ -17,7 +17,7 @@ typedef enum {
 } HAL_PullupConfig;
 
 typedef struct {
-    uint8_t* port;
+    uint16_t* port;
     uint8_t pin;
     HAL_PullupConfig pullup;
     HAL_PinDirection direction;
@@ -28,7 +28,7 @@ typedef struct {
 extern "C" {
 #endif
 
-HAL_Pin* HAL_pinSetup(HAL_Pin* newPin, uint8_t* port, uint8_t pin, HAL_PullupConfig pullup);
+HAL_Pin* HAL_pinSetup(HAL_Pin* newPin, uint16_t* port, uint8_t pin, HAL_PullupConfig pullup);
 void HAL_setPinDirection(HAL_Pin* pin, HAL_PinDirection direction);
 void HAL_pinWrite(HAL_Pin* pin, HAL_PinLevel level);
 HAL_PinLevel HAL_pinRead(HAL_Pin* pin);
