+++ src/I2C_HAL_AVR_Impl.c
@@ -3,7 +3,7 @@
 #include <util/delay.h>
 #include <I2C_HAL.h>
 
-HAL_Pin* HAL_pinSetup(HAL_Pin* newPin, uint8_t* port, uint8_t pin, HAL_PullupConfig pullup){
+HAL_Pin* HAL_pinSetup(HAL_Pin* newPin, uint16_t* port, uint8_t pin, HAL_PullupConfig pullup){
   newPin->port = port;
   newPin->pin = pin;
   newPin->pullup = pullup;
